file upload done

+++ app/Http/Controllers/AttachemntsController.php
@@ -12,11 +12,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 
-class AttachemntsController extends Controller
-{
+class AttachemntsController extends Controller {
 
-    public function __construct()
-    {
+    public function __construct() {
         $this->middleware(['auth']);
     }
 
@@ -25,8 +23,7 @@ class AttachemntsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
+    public function index() {
         $user = Auth::user();
         $pageAuth = $user->authentication(config('auth.privileges.attachments'));
         if ($pageAuth['is_read']) {
@@ -34,8 +31,7 @@ class AttachemntsController extends Controller
         }
     }
 
-    public function isNameUnique($name)
-    {
+    public function isNameUnique($name) {
 
         $user = Auth::user();
         $pageAuth = $user->authentication(config('auth.privileges.attachments'));
@@ -55,8 +51,7 @@ class AttachemntsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
-    {
+    public function create() {
         $user = Auth::user();
         $pageAuth = $user->authentication(config('auth.privileges.attachments'));
         request()->validate([
@@ -83,8 +78,7 @@ class AttachemntsController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store($id)
-    {
+    public function store($id) {
 
         $user = Auth::user();
         $pageAuth = $user->authentication(config('auth.privileges.attachments'));
@@ -92,7 +86,8 @@ class AttachemntsController extends Controller
             'name' => 'required|unique:attachemnts,name',
         ]);
         if ($pageAuth['is_update']) {
-            $attachment = Attachemnt::findOrFail($id);;
+            $attachment = Attachemnt::findOrFail($id);
+            ;
             $attachment->name = \request('name');
             $msg = $attachment->save();
 
@@ -112,8 +107,7 @@ class AttachemntsController extends Controller
      * @param  \App\Attachemnts  $attachemnts
      * @return \Illuminate\Http\Response
      */
-    public function show()
-    {
+    public function show() {
         $user = Auth::user();
         $pageAuth = $user->authentication(config('auth.privileges.attachments'));
         if ($pageAuth['is_read']) {
@@ -123,8 +117,7 @@ class AttachemntsController extends Controller
         }
     }
 
-    public function find($id)
-    {
+    public function find($id) {
 
         $user = Auth::user();
         $pageAuth = $user->authentication(config('auth.privileges.attachments'));
@@ -141,8 +134,7 @@ class AttachemntsController extends Controller
      * @param  \App\Attachemnts  $attachemnts
      * @return \Illuminate\Http\Response
      */
-    public function edit()
-    {
+    public function edit() {
         //
     }
 
@@ -153,8 +145,7 @@ class AttachemntsController extends Controller
      * @param  \App\Attachemnts  $attachemnts
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Attachemnt $attachemnts)
-    {
+    public function update(Request $request, Attachemnt $attachemnts) {
         //
     }
 
@@ -164,12 +155,12 @@ class AttachemntsController extends Controller
      * @param  \App\Attachemnts  $attachemnts
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
+    public function destroy($id) {
         $user = Auth::user();
         $pageAuth = $user->authentication(config('auth.privileges.attachments'));
         if ($pageAuth['is_delete']) {
-            $attachment = Attachemnt::findOrFail($id);;
+            $attachment = Attachemnt::findOrFail($id);
+            ;
             //$attachment->name= \request('name');
             $msg = $attachment->delete();
 
@@ -183,23 +174,22 @@ class AttachemntsController extends Controller
         }
     }
 
-    public function getAttachment_by_application_name($application_name)
-    {
+    public function getAttachment_by_application_name($application_name) {
         $user = Auth::user();
         $pageAuth = $user->authentication(config('auth.privileges.attachments'));
         if ($pageAuth['is_read']) {
             // return Attachemnt::get();
             return Attachemnt::join('application_type_attachemnt', 'application_type_attachemnt.attachemnt_id', '=', 'attachemnts.id')
-                ->join('application_types', 'application_type_attachemnt.application_type_id', '=', 'application_types.id')
-                ->select('attachemnts.*')
-                ->where('application_types.name', '=', $application_name)
-                ->get();
+                            ->join('application_types', 'application_type_attachemnt.application_type_id', '=', 'application_types.id')
+                            ->select('attachemnts.*')
+                            ->where('application_types.name', '=', $application_name)
+                            ->get();
         } else {
             abort(401);
         }
     }
-    public function attach($attachment, $epl)
-    {
+
+    public function attach($attachment, $epl) {
         $user = Auth::user();
         $pageAuth = $user->authentication(config('auth.privileges.attachments'));
         if ($pageAuth['is_create']) {
@@ -208,36 +198,37 @@ class AttachemntsController extends Controller
             $array2 = explode(',', $array[1]);
             $array3 = explode('/', $array[0]);
             $type = $array3[1];
+            if (!($type == 'jpeg' || $type == 'pdf')) {
+//                dd($type);
+                return array('id' => 0, 'message' => 'Only you can add image(jpeg) or PDF');
+            }
             $data = base64_decode($array2[1]);
-            file_put_contents($this->makeApplicationPath($epl, $attachment) . "1" . $type, $data);
-            $path =  $this->makeApplicationPath($epl, $attachment) . "1." . $type;
+            file_put_contents($this->makeApplicationPath($epl, $attachment) . "1." . $type, $data);
+            $path = $this->makeApplicationPath($epl, $attachment) . "1." . $type;
             $e = EPL::findOrFail($epl);
-            return  \DB::transaction(function ()  use ($attachment, $e, $path, $type) {
-                $e->attachemnts()->detach($attachment);
-                $e->attachemnts()->attach(
-                    $attachment,
-                    [
-                        'path' => $path,
-                        'type' => $type,
-                        'created_at' => Carbon::now()->toDateTimeString()
-                    ]
-                );
-                return array('id' => 1, 'message' => 'true');
-            });
+            return \DB::transaction(function () use ($attachment, $e, $path, $type) {
+                        $e->attachemnts()->detach($attachment);
+                        $e->attachemnts()->attach(
+                                $attachment, [
+                            'path' => $path,
+                            'type' => $type,
+                            'created_at' => Carbon::now()->toDateTimeString()
+                                ]
+                        );
+                        return array('id' => 1, 'message' => 'true');
+                    });
         } else {
             abort(401);
         }
     }
 
-    public function revoke($attachment, $epl)
-    {
+    public function revoke($attachment, $epl) {
         $e = EPL::findOrFail($epl);
         $e->attachemnts()->detach($attachment);
         return array('id' => 1, 'message' => 'true');
     }
 
-    public function getEplAttachments($epl)
-    {
+    public function getEplAttachments($epl) {
         return \DB::select(\DB::raw("SELECT b.path, b.type, b.attachment_epl_id, a.att_id, a.attachment_name FROM (SELECT
 	application_types.`name`, 
 	attachemnts.`name` AS attachment_name,
@@ -266,8 +257,7 @@ WHERE attachemnt_e_p_l.e_p_l_id = '{$epl}') AS b
         //        return $attachemntsAssigned = $epl->attachemnts;
     }
 
-    private function makeApplicationPath($id, $attachemntId)
-    {
+    private function makeApplicationPath($id, $attachemntId) {
         if (!is_dir("uploads")) {
             //Create our directory if it does not exist
             mkdir("uploads");
@@ -290,4 +280,5 @@ WHERE attachemnt_e_p_l.e_p_l_id = '{$epl}') AS b
         }
         return "uploads/EPL/" . $id . "/attachments/" . $attachemntId . "/";
     }
+
 }
